requirements.txt
# ============================================
# Analizador de Documentos Legales - Dependencies
# ============================================
# Python 3.10+ required

# UI Framework
streamlit>=1.30.0

# Text Extraction - PDF
pdfplumber>=0.10.0

# Text Extraction - DOCX
python-docx>=1.1.0

# Image Processing and OCR
pillow>=10.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10

# HTTP Client for Ollama
requests>=2.31.0

# Data Validation and Schemas
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development Tools (optional)
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# Logging and Configuration
pyyaml>=6.0.1
